requirements.txt
gitpython==3.1.41
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.3.0
rich==13.7.0
click==8.1.7 